Make tests recognizable to pytest
